# backend/api/openapi/tracking.yaml
openapi: 3.0.3
info:
  title: defShows Tracking API
  version: 0.1.0
security:
  - bearerAuth: []
paths:
  /me/shows:
    get:
      operationId: listTracked
      summary: List the user's tracked shows with progress
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackedShowList' }
        '401': { $ref: '#/components/responses/Error' }
    post:
      operationId: addShow
      summary: Track a show (imports from TMDB on first use)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddShowRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserShow' }
        '409': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /me/shows/{showId}:
    get:
      operationId: getTracked
      summary: One tracked show with progress
      parameters: [{ $ref: '#/components/parameters/ShowId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackedShow' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateShow
      summary: Update status/favorite/dubbing
      parameters: [{ $ref: '#/components/parameters/ShowId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateShowRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserShow' }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      operationId: removeShow
      summary: Stop tracking a show
      parameters: [{ $ref: '#/components/parameters/ShowId' }]
      responses:
        '204': { description: No Content }
  /me/shows/{showId}/watch:
    post:
      operationId: watchShow
      summary: Mark all catalogued episodes watched and set tracking status to completed
      description: Atomic and idempotent. Preserves existing episode ratings and watch dates. Episodes added to the catalog later are not automatically watched.
      parameters: [{ $ref: '#/components/parameters/ShowId' }]
      responses:
        '204': { description: No Content }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /me/shows/{showId}/episodes/{episodeId}/watch:
    post:
      operationId: watchEpisode
      summary: Mark an episode watched
      parameters:
        - { $ref: '#/components/parameters/ShowId' }
        - { $ref: '#/components/parameters/EpisodeId' }
      responses:
        '204': { description: No Content }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      operationId: unwatchEpisode
      summary: Mark an episode unwatched
      parameters:
        - { $ref: '#/components/parameters/ShowId' }
        - { $ref: '#/components/parameters/EpisodeId' }
      responses:
        '204': { description: No Content }
        '404': { $ref: '#/components/responses/Error' }
  /me/shows/{showId}/links:
    get:
      operationId: listLinks
      summary: List reference links
      parameters: [{ $ref: '#/components/parameters/ShowId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LinkList' }
        '404': { $ref: '#/components/responses/Error' }
    post:
      operationId: addLink
      summary: Add a reference link
      parameters: [{ $ref: '#/components/parameters/ShowId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddLinkRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Link' }
        '404': { $ref: '#/components/responses/Error' }
  /me/shows/{showId}/links/{linkId}:
    patch:
      operationId: updateLink
      summary: Update a reference link value (URL or plain text)
      parameters:
        - { $ref: '#/components/parameters/ShowId' }
        - name: linkId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateLinkRequest' }
      responses:
        '204': { description: No Content }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteLink
      summary: Delete a reference link
      parameters:
        - { $ref: '#/components/parameters/ShowId' }
        - name: linkId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: No Content }
  /me/recommendations:
    get:
      operationId: getRecommendations
      summary: Taste profile and catalog recommendations for the home dashboard
      description: Derived from the user's tracked shows over the local catalog; cached per user with a TTL.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HomeRecommendations' }
        '401': { $ref: '#/components/responses/Error' }
  /me/stats:
    get:
      operationId: getStats
      summary: Aggregate viewing statistics for the current user
      description: Counts of watched shows/seasons/episodes, total watch time, and how many shows are tracked in total.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Stats' }
        '401': { $ref: '#/components/responses/Error' }
  /me/settings:
    get:
      operationId: getSettings
      summary: Get user settings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Settings' }
    patch:
      operationId: updateSettings
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSettingsRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Settings' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    ShowId:
      name: showId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    EpisodeId:
      name: episodeId
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
    AddShowRequest:
      type: object
      required: [tmdb_id]
      properties:
        tmdb_id: { type: integer, format: int64 }
    UserShow:
      type: object
      required: [id, show_id, status, favorite]
      properties:
        id: { type: integer, format: int64 }
        show_id: { type: integer, format: int64 }
        status: { type: string, enum: [watching, plan_to_watch, on_hold, completed, dropped] }
        favorite: { type: boolean }
        preferred_dubbing: { type: string }
        rating: { type: integer, minimum: 1, maximum: 10, nullable: true, description: The user's own 1..10 score; null when unrated }
    ShowRef:
      type: object
      required: [id, tmdb_id, title, airing_status]
      properties:
        season_count: { type: integer, minimum: 0, description: Catalog seasons excluding specials }
        id: { type: integer, format: int64 }
        tmdb_id: { type: integer, format: int64 }
        title: { type: string }
        original_title: { type: string }
        poster_url: { type: string }
        airing_status: { type: string, enum: [not_started, airing, between_seasons, ended] }
        next_episode_air_date: { type: string, nullable: true }
        vote_average: { type: number }
        vote_count: { type: integer, format: int64 }
    Progress:
      type: object
      required: [watched, total, unwatched, watched_episode_ids]
      properties:
        watched_episode_ids:
          type: array
          items: { type: integer, format: int64 }
        watched: { type: integer, description: Watched episodes that have already aired }
        total: { type: integer, description: Episodes with a known air date of today or earlier }
        unwatched: { type: integer, description: Aired but not-yet-watched episodes }
        next_unwatched_episode_id: { type: integer, format: int64, nullable: true }
        new_full_season:
          type: integer
          nullable: true
          description: >-
            Season number of a fully-aired, completely-unwatched season that dropped while
            the user was caught up on everything earlier; null when there is no such season.
    TrackedShow:
      type: object
      required: [user_show, show, progress]
      properties:
        user_show: { $ref: '#/components/schemas/UserShow' }
        show: { $ref: '#/components/schemas/ShowRef' }
        progress: { $ref: '#/components/schemas/Progress' }
    TrackedShowList:
      type: object
      required: [tracked]
      properties:
        tracked:
          type: array
          items: { $ref: '#/components/schemas/TrackedShow' }
    UpdateShowRequest:
      type: object
      properties:
        status: { type: string, enum: [watching, plan_to_watch, on_hold, completed, dropped] }
        favorite: { type: boolean }
        preferred_dubbing: { type: string }
        rating: { type: integer, minimum: 1, maximum: 10, description: Set the user's 1..10 score }
        clear_rating: { type: boolean, description: "When true, remove the existing rating (ignored if rating is also set)" }
    Link:
      type: object
      required: [id, kind, url]
      properties:
        id: { type: integer, format: int64 }
        kind: { type: string, enum: [download, streaming, wiki, imdb, kinopoisk] }
        label: { type: string }
        url: { type: string }
    UpdateLinkRequest:
      type: object
      required: [url]
      properties:
        url: { type: string }
    AddLinkRequest:
      type: object
      required: [kind, url]
      properties:
        kind: { type: string, enum: [download, streaming, wiki, imdb, kinopoisk] }
        label: { type: string }
        url: { type: string }
    LinkList:
      type: object
      required: [links]
      properties:
        links:
          type: array
          items: { $ref: '#/components/schemas/Link' }
    TasteGenre:
      type: object
      required: [id, name, weight]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        weight: { type: integer }
    TasteLanguage:
      type: object
      required: [code, weight]
      properties:
        code: { type: string }
        weight: { type: integer }
    Taste:
      type: object
      required: [genres, languages]
      properties:
        genres:
          type: array
          items: { $ref: '#/components/schemas/TasteGenre' }
        languages:
          type: array
          items: { $ref: '#/components/schemas/TasteLanguage' }
    HomeRecommendations:
      type: object
      required: [taste, recommendations]
      properties:
        taste: { $ref: '#/components/schemas/Taste' }
        recommendations:
          type: array
          items: { $ref: '#/components/schemas/ShowRef' }
    Stats:
      type: object
      required: [shows_tracked, shows_completed, seasons_watched, episodes_watched, minutes_watched]
      properties:
        shows_tracked: { type: integer, description: Total shows the user tracks }
        shows_completed: { type: integer, description: Shows with status completed }
        seasons_watched: { type: integer, description: Seasons whose aired episodes are all watched }
        episodes_watched: { type: integer, description: Episodes marked watched }
        minutes_watched: { type: integer, format: int64, description: "Summed runtime of watched episodes, in minutes" }
    Settings:
      type: object
      required: [timezone, is_public]
      properties:
        timezone: { type: string }
        is_public: { type: boolean }
    UpdateSettingsRequest:
      type: object
      required: [timezone]
      properties:
        timezone: { type: string }
        is_public: { type: boolean }